requirements.txt
requests==2.25.1
Flask==1.1.2
flask-restful~=0.3.8
flask-cors~=2.1.3
python-dotenv~=0.17.1
beautifulsoup4~=4.9.3
lxml~=4.6.3